Changed peragro specific config file keys to start with "Peragro."; this is consistent with the other crystal space applications. Sorry, your custom configurations will be broken; check data/config/client.cfg for the new format.

+++ plugins/terrained/terrained.cpp
@@ -450,12 +450,12 @@ bool TerrainEd::OnShaderSpinner(const CEGUI::EventArgs& e)
   csArray<float> vals = GetValues(win);
 
   csShaderVariable::VariableType type = var->GetType();
-  printf("I: Setting var of type %d  len %d!\n", type, vals.GetSize());
+  printf("I: Setting var of type %d  len %lu!\n", type, vals.GetSize());
   switch (type)
   {
   case csShaderVariable::INT : 
     {
-      int val = vals[0];
+      int val = static_cast<int>(vals[0]);
       var->SetValue(val);
     }
     break;
